requirements.txt
# secIRC - Anonymous Censorship-Resistant Messaging System
# Python Dependencies

# Core Framework
pydantic>=2.5.0
pydantic-settings>=2.1.0
asyncio>=3.4.3

# Network & UDP Communication
asyncio-dgram>=2.1.2
aiofiles>=23.2.0
dnspython>=2.4.0

# Security & Cryptography (Enhanced for Anonymous Messaging)
cryptography>=41.0.0
pycryptodome>=3.19.0
argon2-cffi>=23.1.0
PyNaCl>=1.5.0  # PyNaCl for modern crypto

# Anonymous Network & Discovery
requests>=2.31.0
aiohttp>=3.9.0
beautifulsoup4>=4.12.0  # For web scraping relay discovery

# Key Management & Storage
keyring>=24.3.0

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.7.0

# Logging & Monitoring (Minimal for anonymity)
structlog>=23.2.0

# Utilities
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.7.0
typer>=0.9.0

# Network Security & Anonymity
scapy>=2.5.0  # For network packet analysis

# Torrent Protocol Support
bencodepy>=2.0.0  # For BitTorrent bencode support
